# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Recipes API
# during dev, should point to your local machine
host: localhost:10010
# basePath prefixes all resource paths 
basePath: /v1  
# 
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
  - text/plain

securityDefinitions:
  JWTAuth:
    type: apikey
    name: Authorization
    in: header
  
paths:
  /recipes:
    x-swagger-router-controller: recipe
    get:
      description: Return all the recipes
      operationId: getRecipes
      responses: 
        "200":
          description: Sucess get all the recipes
          schema:
            type: array
            items:
              $ref: "#/definitions/Recipe"
        "500":
          description: Unexpected Error
          schema:
            type: object
            properties:
              message:
                type: string
    post:
      description: Create one new Recipe
      operationId: postRecipes
      parameters:
        - in: body
          name: body
          description: The recipe to be added
          required: true
          schema:
            $ref: "#/definitions/Recipe"
      responses:
        "204":
          description: Sucess
        "500":
          description: unexpected Error
          schema:
            type: object
            properties:
              message:
                type: string
      
  /recipes/{recipeId}:
    x-swagger-router-controller: recipe
    parameters:
    - in: path
      name: recipeId
      type: integer
      required: true
      description: Numeric Id of the recipe
    delete:
      description: Remove a recipe
      operationId: deleteRecipe
      responses:
        "204":
          description: Success
        "404":
          description: The Id passed does not match any recipe.
        "500":
          description: unexpected Error
          schema:
            type: object
            properties:
                message:
                    type: string
    put:
      parameters:
      - in: body
        name: Recipe
        schema:
          $ref: "#/definitions/Recipe"
        required: true
        description: Id of the recipe you want to get the comments.
      description: updates a Recipe
      operationId: putRecipes
      responses:
        "204":
          description: Success
        "404":
          description: The Id passed does not match any recipe.
        "500":
          description: unexpected Error
          schema:
            type: object
            properties:
                message:
                    type: string

  /recipes/{recipeId}/comments:
    x-swagger-router-controller: comments
    parameters:
    - in: path
      name: recipeId
      type: integer
      required: true
      description: Numeric Id of the recipe
    get:
      description: get all the comments
      operationId: getComments
      responses:
        "200":
          description: Success
          schema:
            type: array
            items:
              $ref: "#/definitions/Comment"
        "404":
          description: The Recipe Id is not a valid one.
        "500":
          description: unexpected Error
          schema:
            type: object
            properties:
                message:
                    type: string
              
    post:
      parameters:
        - in: body
          name: comment
          schema:
            $ref: "#/definitions/Comment"
          required: true
      operationId: postComments
      responses:
        "204":
          description: Success
        "404":
          description: The Recipe Id is not a valid one.
        "500":
          description: unexpected Error
          schema:
            type: object
            properties:
                message:
                    type: string
                    
  /recipes/{recipeId}/comments/{commentId}:
    x-swagger-router-controller: comments
    parameters:
      - in: path
        name: recipeId
        type: integer
        required: true
        description: Id of the recipe you want to delete the comments.
      - in: path
        name: commentId
        type: integer
        required: true
        description: Id of the comment which you want to delete.
    delete:
      description: deletes one comment
      operationId: deleteComments
      responses:
        "204":
          description: Success the comment is eliminated
        "404":
          description: The Recipe or the comment was not found.
        "500":
          description: unexpected Error
          schema:
            type: object
            properties:
                message:
                    type: string 
          
  /swagger:
    x-swagger-pipe: swagger_raw
    
#  /login:
#    x-swagger-router-controller: auth
#    post:
#      tags:
#        - authetication
#      operationId: login
#      parameters:
#      - name: body
#        in: body
#        schema:
#          properties:
#            email:
#              type: string
#            password:
#              type: string
#      responses:
#        "200":
#          description: Success
#          schema: 
#            properties:
#              token:
#                type: string
#        "400":
#          description: Invalid Login
#          schema: 
#            $ref: "#/definitions/Erroresponse"
            
#  /secret-recipe:
#    x-swagger-router-controller: recipe
#    get:
#      description: Gets the most amazing and secret chefs recipe
#      security:
#        -JWTAuth: []
#        operationId: getSecretRecipe
#      responses:
#        "200":
#          description: Success you get the amaxzing chefs recipe
#          schema: 
#            $ref: "#/definitions/Recipe"
#        "400":
#          description: Invalid Login
#          schema: 
#            $ref: "#/definitions/Errorresponse"
            
        
    
# complex objects have schema definitions
definitions:
  Recipe:
    properties:
      id:
        type: integer
      isPublic:
        type: boolean
      name: 
        type: string
      chef:
        type: string
        enum: [Joao, Charles]
      stepList:
        type: array
        items:
          type: string
      ingredients:
        type: array
        items:
          type: object
          properties:
            amount:
              type: number
            name:
              type: string
              
  Comment:
    properties:
      id:
        type: integer
      recipeId:
        type: integer
      author:
        type: string
      text:
        type: string
      date:
        type: string